HW-4 BaseActivity 추가 및 상속구조 변경
