+++ jskim5923/ArchitectureStudy/app/src/main/java/com/jskim5923/architecturestudy/base/BaseActivity.kt
@@ -1,0 +1,25 @@
+package com.jskim5923.architecturestudy.base
+
+import android.os.Bundle
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+
+abstract class BaseActivity : AppCompatActivity(), BaseContract.View {
+    @get:LayoutRes
+    abstract val layoutRes: Int
+
+    abstract val presenter: BaseContract.Presenter
+
+    abstract override fun initView()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(layoutRes)
+        initView()
+    }
+
+    override fun onDestroy() {
+        presenter.clearCompositeDisposable()
+        super.onDestroy()
+    }
+}
