+++ jskim5923/ArchitectureStudy/app/src/main/java/com/jskim5923/architecturestudy/main/MainActivity.kt
@@ -1,38 +1,30 @@
 package com.jskim5923.architecturestudy.ui
 
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
 import com.jskim5923.architecturestudy.R
 import com.jskim5923.architecturestudy.adapter.ViewPagerAdapter
+import com.jskim5923.architecturestudy.base.BaseActivity
 import com.jskim5923.architecturestudy.main.MainContract
 import com.jskim5923.architecturestudy.main.MainPresenter
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity(), MainContract.View {
+class MainActivity : BaseActivity(), MainContract.View {
+    override val layoutRes = R.layout.activity_main
+
     private val viewpagerAdapter by lazy {
         ViewPagerAdapter(supportFragmentManager)
     }
 
-    private val presenter = MainPresenter(this)
+    override val presenter = MainPresenter(this)
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
+    override fun initView() {
         with(viewPager) {
             tabLayout.setupWithViewPager(this)
             adapter = viewpagerAdapter
         }
-
         presenter.loadMarketList()
     }
 
     override fun updateViewpager(marketList: List<String>) {
         viewpagerAdapter.updateItem(marketList)
-    }
-
-    override fun onDestroy() {
-        presenter.clearCompositeDisposable()
-        super.onDestroy()
     }
 }
